cache unhandled slack events in redis for later examination
Currently Slack sends us two events: `message.channels` and `team_join`.
We process the former already, but not yet the latter. It's not clear to
me what the payload looks like so this patch adds support to capture and
cache the last 100 unhandled slack events sent to the app so we can
inspect them and implement them later.