add code of conduct
fixes #83

This patch puts the CoC in a markdown file in the root of the repository
and adds a rails route to load it in, convert it to HTML, and display it.

It also puts a link to the CoC in the footer, replacing twitter which we
already had an icon link to.