send all slack events on railslink-dev.heroku.com
I need to be able to see the payloads for any slack events in order to
facilite some future development. Instead of stealing the web hooks for
railslink-dev and setting up a tunnel this patch simply sends all slack
events through the existing unhandled workflow which will record the
last 100 events.