remove is_heroku_dev_app check
It turns out we don't need this because we already do redis publishing
of all events and I just need to subscribe to it to see it. I even wrote
that code, but forgot about it.

So this just removes something that I never should have added in the
first place.